Merged and fixed errors
Should be all good....hopefully

Also as a reminder if you are reading this, please stop deleting things, like in general. Add a comment or send a message in the discord if you think something is arbitrary or useless, because sometimes it's not and that leads to a lot lot of errors for other people.